+++ src/bindings/vips-operators.h
@@ -520,7 +520,15 @@ static Image pngload_source(const Source &source, emscripten::val js_options = e
 static Image ppmload(const std::string &filename, emscripten::val js_options = emscripten::val::null());
 
 /**
- * Load ppm base class.
+ * Load ppm from buffer.
+ * @param buffer Buffer to load from.
+ * @param js_options Optional options.
+ * @return Output image.
+ */
+static Image ppmload_buffer(const std::string &buffer, emscripten::val js_options = emscripten::val::null());
+
+/**
+ * Load ppm from source.
  * @param source Source to load from.
  * @param js_options Optional options.
  * @return Output image.
@@ -1739,10 +1747,17 @@ Image math(emscripten::val math) const;
 Image math2_const(emscripten::val math2, const std::vector<double> &c) const;
 
 /**
- * Invert an matrix.
+ * Invert a matrix.
  * @return Output matrix.
  */
 Image matrixinvert() const;
+
+/**
+ * Multiply two matrices.
+ * @param right Second matrix to multiply.
+ * @return Output matrix.
+ */
+Image matrixmultiply(emscripten::val right) const;
 
 /**
  * Print matrix.
@@ -2053,6 +2068,14 @@ Image relational_const(emscripten::val relational, const std::vector<double> &c)
 Image remainder_const(const std::vector<double> &c) const;
 
 /**
+ * Rebuild an mosaiced image.
+ * @param old_str Search for this string.
+ * @param new_str And swap for this string.
+ * @return Output image.
+ */
+Image remosaic(const std::string &old_str, const std::string &new_str) const;
+
+/**
  * Replicate an image.
  * @param across Repeat this many times horizontally.
  * @param down Repeat this many times vertically.
@@ -2123,7 +2146,7 @@ Image scRGB2BW(emscripten::val js_options = emscripten::val::null()) const;
 Image scRGB2XYZ() const;
 
 /**
- * Convert an scrgb image to srgb.
+ * Convert scrgb to srgb.
  * @param js_options Optional options.
  * @return Output image.
  */