+++ src/vips-emscripten.cpp
@@ -149,12 +149,6 @@ EMSCRIPTEN_BINDINGS(my_module) {
         .value("scrgb", VIPS_INTERPRETATION_scRGB)
         .value("hsv", VIPS_INTERPRETATION_HSV);
 
-    enum_<VipsDemandStyle>("DemandStyle")
-        .value("error", VIPS_DEMAND_STYLE_ERROR)
-        .value("smalltile", VIPS_DEMAND_STYLE_SMALLTILE)
-        .value("fatstrip", VIPS_DEMAND_STYLE_FATSTRIP)
-        .value("thinstrip", VIPS_DEMAND_STYLE_THINSTRIP);
-
     enum_<VipsOperationRelational>("OperationRelational")
         .value("equal", VIPS_OPERATION_RELATIONAL_EQUAL)
         .value("noteq", VIPS_OPERATION_RELATIONAL_NOTEQ)
@@ -384,7 +378,8 @@ EMSCRIPTEN_BINDINGS(my_module) {
         .value("perceptual", VIPS_INTENT_PERCEPTUAL)
         .value("relative", VIPS_INTENT_RELATIVE)
         .value("saturation", VIPS_INTENT_SATURATION)
-        .value("absolute", VIPS_INTENT_ABSOLUTE);
+        .value("absolute", VIPS_INTENT_ABSOLUTE)
+        .value("auto", VIPS_INTENT_AUTO);
 
     enum_<VipsKernel>("Kernel")
         .value("nearest", VIPS_KERNEL_NEAREST)
@@ -392,7 +387,9 @@ EMSCRIPTEN_BINDINGS(my_module) {
         .value("cubic", VIPS_KERNEL_CUBIC)
         .value("mitchell", VIPS_KERNEL_MITCHELL)
         .value("lanczos2", VIPS_KERNEL_LANCZOS2)
-        .value("lanczos3", VIPS_KERNEL_LANCZOS3);
+        .value("lanczos3", VIPS_KERNEL_LANCZOS3)
+        .value("mks2013", VIPS_KERNEL_MKS2013)
+        .value("mks2021", VIPS_KERNEL_MKS2021);
 
     enum_<VipsPCS>("PCS")
         .value("lab", VIPS_PCS_LAB)
@@ -1348,6 +1345,10 @@ EMSCRIPTEN_BINDINGS(my_module) {
         .class_function("ppmload", optional_override([](const std::string &filename) {
                             return Image::ppmload(filename);
                         }))
+        .class_function("ppmloadBuffer", &Image::ppmload_buffer)
+        .class_function("ppmloadBuffer", optional_override([](const std::string &buffer) {
+                            return Image::ppmload_buffer(buffer);
+                        }))
         .class_function("ppmloadSource", &Image::ppmload_source)
         .class_function("ppmloadSource", optional_override([](const Source &source) {
                             return Image::ppmload_source(source);
@@ -1807,6 +1808,7 @@ EMSCRIPTEN_BINDINGS(my_module) {
                       return image.match(sec, xr1, yr1, xs1, ys1, xr2, yr2, xs2, ys2);
                   }))
         .function("matrixinvert", &Image::matrixinvert)
+        .function("matrixmultiply", &Image::matrixmultiply)
         .function("matrixprint", &Image::matrixprint)
         .function("matrixprint", optional_override([](const Image &image) {
                       image.matrixprint();
@@ -1923,6 +1925,7 @@ EMSCRIPTEN_BINDINGS(my_module) {
         .function("reducev", optional_override([](const Image &image, double vshrink) {
                       return image.reducev(vshrink);
                   }))
+        .function("remosaic", &Image::remosaic)
         .function("replicate", &Image::replicate)
         .function("resize", &Image::resize)
         .function("resize", optional_override([](const Image &image, double scale) {